Force password reset on next login after password is manually set
Squashed commit of the following:

commit a79ddf21836e89853cb57be9da4d5e0619892bb1
Author: drlippman <drlippman@yahoo.com>
Date:   Fri Dec 29 14:30:46 2017 -0800

    Force cache reload validate JS

commit 3277cb63c869e9107504e78cbe8fb224dbf9877c
Author: drlippman <drlippman@yahoo.com>
Date:   Fri Dec 29 14:27:25 2017 -0800

    Force PW reset on batchcreateinstr

commit 2692d8acdec763b8a6b8fec24b8c9c6bf32f1449
Author: drlippman <drlippman@yahoo.com>
Date:   Fri Dec 29 14:23:04 2017 -0800

    Change language on PW reset to reflect PW is temporary

commit 61ee2327bff6598873c61afa7b817a9802e8d320
Author: drlippman <drlippman@yahoo.com>
Date:   Wed Dec 27 12:59:26 2017 -0800

    Feature to force password reset on first login after manual reset

+++ forms.php
@@ -24,7 +24,7 @@ if (isset($_GET['greybox'])) {
 } else {
 	$gb = '';
 }
-$placeinhead = '<script type="text/javascript" src="'.$imasroot.'/javascript/jquery.validate.min.js"></script>';
+$placeinhead = '<script type="text/javascript" src="'.$imasroot.'/javascript/jquery.validate.min.js?v=122917"></script>';
 if (isset($CFG['locale'])) {
 	$placeinhead .= '<script type="text/javascript" src="'.$imasroot.'/javascript/jqvalidatei18n/messages_'.$CFG['locale'].'.min.js"></script>';
 }
@@ -94,12 +94,18 @@ switch($_GET['action']) {
 			include($studentTOS);
 		}
 		break;
+	case "forcechgpwd":
 	case "chgpwd":
-		if ($gb == '') {
+		if ($gb == '' && $_GET['action']!='forcechgpwd') {
 			echo "<div class=breadcrumb><a href=\"index.php\">Home</a> &gt; Change Password</div>\n";
 		}
 		echo '<div id="headerforms" class="pagetitle"><h2>Change Your Password</h2></div>';
-		echo "<form id=\"pageform\" class=limitaftervalidate method=post action=\"actions.php?action=chgpwd$gb\">\n";
+		if ($_GET['action']=='forcechgpwd') {
+			echo '<p>'._('To ensure the security of your account, we are requiring a password change. Please select a new password.').'</p>';
+			echo "<form id=\"pageform\" class=limitaftervalidate method=post action=\"actions.php?action=forcechgpwd$gb\">\n";
+		} else {
+			echo "<form id=\"pageform\" class=limitaftervalidate method=post action=\"actions.php?action=chgpwd$gb\">\n";
+		}
 		echo "<span class=form><label for=\"oldpw\">Enter old password:</label></span> <input class=form type=password id=oldpw name=oldpw size=40 /> <BR class=form>\n";
 		echo "<span class=form><label for=\"pw1\">Enter new password:</label></span>  <input class=form type=password id=pw1 name=pw1 size=40> <BR class=form>\n";
 		echo "<span class=form><label for=\"pw2\">Verify new password:</label></span>  <input class=form type=password id=pw2 name=pw2 size=40> <BR class=form>\n";
